+++ sailbrowser-web/src/app/results-input/scoring-sheet-scanner/run-scan/scan-execution.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { vi } from 'vitest';
 import { FirebaseApp } from '@angular/fire/app';
 import { ScanExecutionService } from './scan-execution.service';
-import { ScanRunState } from '../scan-model';
+import { ScanRunState } from '../model/scan-model';
 
 describe('ScanExecutionService', () => {
   let service: ScanExecutionService;
